+++ CHANGELOG.md
@@ -11,6 +11,7 @@ _Unreleased_
 * Correct weekday order and default holidays for various countries.
 * Now warning about `amenity=fixme` objects.
 * Untyped amenities with just a name will be converted to OSM notes.
+* "Clear" button for the fixme field (thanks @soham-dixit).
 * Added `sr-Latn` language for Serbia (thanks @alnzrv).
 * Updated roof assets (thanks @mxdanger).
 * Updated a phone number parsing library.